copmleted 2.18 + moved components into own jsx file
