+++ part2/countries/src/components/SearchFilter.jsx
@@ -1,31 +1,16 @@
-// const Filter = ({ countries, country, handleFilterChange }) => {
-    
-//     const returnDisplay =  () => {   return (
-//         <>
-//             filter countries: <input value={country} onChange={handleFilterChange}/>
-//             {filteredCountries.map(country => {
-//                 <div>{country.name.common}</div>
-//             })}
-//         </>
-//     )}
+import Display from "./DisplayCountry"
 
-//     const filterMatches = countries.filter(country => country.name.common.toLowerCase().startsWith(country))
+const Filter = ({ countries, showCountry, handleFilterChange }) => {
+    const filterMatches = countries.filter(country => 
+      country.name.common.toLowerCase().startsWith(showCountry.toLowerCase()))
+  
+    return (
+      <>
+        search country: <input value={showCountry} onChange={handleFilterChange}/>
+        <Display countries={filterMatches} showCountry={showCountry}/>
+      </>
+    )
+  }
+  
 
-//     if (filterMatches.length > 10 && country != '') {
-//         const filteredCountries = ['too many results...']
-//         return returnDisplay
-        
-
-//     } else if (filterMatches.length > 1 && filterMatches.length < 10) {
-//         const filteredCountries = filterMatches
-//         return returnDisplay
-
-//     } else {
-//         const filteredCountries=filterMatches[0]
-//         return returnDisplay
-//     }
-
-
-// }
-
-// export default Filter
+export default Filter
